// src/codegen/wasm/types.ts
export const enum WasmType {
    i32 = 0x7f,
    i64 = 0x7e,
    f32 = 0x7d,
    f64 = 0x7c,
    v128 = 0x7b,
    funcref = 0x70,
    externref = 0x6f,
    void = 0x40,  // empty block type
}